hero::GetLuck +0.32 on a long/int clamp temp, and two string-describer levers measured
GetLuck 87.7439 -> 88.0662: `luck` is `long`, not `int`.  `_cpp_clamp` is
declared `const int&(const int&, const int&, const int&)`, so a `long`
lvalue needs a conversion and therefore a THIRD temporary.  Retail
materialises three (-3 at [ebp+8], 3 at [ebp+0xc], the value at
[ebp+0x10]); the `int` spelling binds `luck`'s own home for the second
argument, needs only two temps, and leaves two dead `mov [ebp+0xc], eax`
write-backs retail does not have.  The whole tail now agrees on slot
assignment and the body is 12 bytes shorter.

Residual notes for get_morale_description / get_luck_description record a
MEASURED but UNSHIPPABLE route: retail's PBDI-append census is 18 calls +
3 expansions (morale) and 12 + 3 (luck) against our 17 + 4 and 11 + 4, and
the surplus expansions sit at named rungs.  Pinning them plus respelling
the two tail `+=` sites as `append(format_string(...), 0, npos)` reaches
93.91 and 95.45 - but it needs six new `#pragma inline_depth(0)` pins and
the cleanliness ratchet holds pins at 355 falling-only.  The same call-site
spellings WITHOUT the pins score 76.26, so the pin is load-bearing.  Both
rows' MAX now sits at the measured peak; the notes say exactly how to
reach it if the pin floor is ever lifted.

+++ src/hero.cpp
@@ -3649,6 +3649,28 @@ std::string hero::get_morale_description() const
         morale++;
     }
     if (flags & 0x200000) {
+        // Residual (91.65%) - MEASURED, BLOCKED BY THE PIN FLOOR
+        // (polish 26).  Retail CALLS append(const char*,size_type) at THIS
+        // rung too (0x4dc71b `repne scasb` + call, a 0x20-byte arm against
+        // our 0x73-byte expansion); the fn-level census is retail 18
+        // PBDI-append calls + 3 _Grow/_Xlen/_Eos expansions against our 17
+        // + 4, and this is the surplus expansion.  Spelling it
+        // `const char* t = gMoraleTexts[19]; result.append(t, strlen(t));`
+        // under a statement `inline_depth(0)` moves the first divergence
+        // from +0x408 to +0x551 and aligns two of the three _Grow sites
+        // exactly - but the freed budget then over-inlines the Grail and
+        // negative-modifier `+=` sites below, so the pin only pays PAIRED
+        // with pins on those two spelled as
+        // `result.append(format_string(...), 0, std::string::npos)`:
+        // 91.65 -> 89.97 (this rung alone) -> 93.21 (+ both `+=` pins)
+        // -> 93.91 (+ the two tail sites respelled as append).  All three
+        // pins are needed together; the same three call-site spellings
+        // WITHOUT the pins score 76.26.  Not shippable while the
+        // cleanliness ratchet holds inline-depth pins at 355 falling-only.
+        // Also rejected: a named `std::string grailText` local for the
+        // Grail temp gives a PERFECT skeleton (109/110 blocks, 0 missing,
+        // 88 exact) and only 92.16 - the extra frame slot costs the tail's
+        // register allocation.
         result += gMoraleTexts[19];
         morale -= 3;
     }
@@ -3808,6 +3830,18 @@ std::string hero::get_luck_description() const
     }
 
     if (skillLevel[eSecSkillLuck] == eMasteryBasic) {
+        // Residual (93.71%) - MEASURED, BLOCKED BY THE PIN FLOOR
+        // (polish 26).  Identical to the morale twin: retail's PBDI-append
+        // census is 12 calls + 3 expansions against our 11 + 4, and the
+        // surplus expansion sits at THIS rung and at the [13] siren rung
+        // above - pinning either alone just moves the flip to the other.
+        // Pinning BOTH (`const char* t = gLuckTexts[n];
+        // result.append(t, strlen(t));` under `inline_depth(0)`) reaches
+        // 92.15, and adding the negative-modifier arm respelled as
+        // `result.append(format_string(...), 0, std::string::npos)` under
+        // the same pin reaches 95.45 (+1.74 over the banked max).
+        // Three new pins; not shippable while the ratchet holds pins at
+        // 355 falling-only.
         result += gLuckTexts[15];
         luck++;
     }
@@ -6648,7 +6682,14 @@ int hero::GetLuck(const hero* otherHero, unsigned char on_cursed_ground,
             return 0;
     }
 
-    int luck = kLuckBonuses[skillLevel[eSecSkillLuck]];
+    // `long`, not `int`, and it is byte-proven: `_cpp_clamp` takes
+    // `const int&`, so a `long` lvalue needs a CONVERSION and therefore a
+    // third temporary.  Retail materialises three (-3 at [ebp+8], 3 at
+    // [ebp+0xc] and the value at [ebp+0x10]) where an `int` binds the
+    // variable's own home directly and only needs two; the `int` spelling
+    // also left two dead `mov [ebp+0xc], eax` write-backs of `luck` that
+    // retail does not have.  87.7439 -> 88.0662.
+    long luck = kLuckBonuses[skillLevel[eSecSkillLuck]];
     if (skillLevel[eSecSkillLuck] > 0) {
         const THeroSpecificAbility& ability = akHeroSpecificAbilities[id];
         if (ability.type == eHeroAbilitySecondarySkill &&
